Move openHAB settings to it's own page
This also changes, how the settings looks like as a preparation to
add more than one openHAB instance. The one possible openHAB instance
is now displayed in a table, which could hold more instances, too.

The update functionality is still the same and updates the one instance,
instead of adding a new one, this will follow in next commits.