Add possibility to add multiple openHAB instances
Instead of just having one openHAB instance per account, openHAB-cloud
now allows multiple instances to be added to the account. However,
most functions still only uses one openHAB instance, which is returned
by findOne (the order of openHAB instances returned is not guaranteed.

A warning is added, so that the user knows, that this feature is
experimental. Also, the feature is behind a feature flag, which needs
to be explicitly enabled in the config.json.