Implement identification of proxied requests
Until now, the openHAB requests proxied to the openHAB instance,
which was found by the findOne() method of mongoose, except for the
case when an UUID was passed as an URL parameter.

Now, if an UUID is passed as an URL parameter, the UUID will also be
send as the openhab-uuid cookie, which allows future requests to be
identified to be made against this specific openHAB instance.