Merge sitemaps of all openHAB instances for /rest/sitemaps
Instead of proxying the /rest/sitemaps api endpoint directly to the
first openHAB instance returned for the user, openHAB-cloud now handles
this REST api endpoint different if multi-instance mode is enabled.

If the mode is enabled, openHAB-cloud will get all openHAB instances
registered for the user account and issues a request to /rest/sitemaps
for each openHAB instance. The result is then merged into a merged array
of all sitemaps. The links in the sitemaps are also rewritten to include
the openHAB uuid in it, so openHAB-cloud can identify requests to belong
to a specific openHAB instance. The resulting array will be returned to
the user / app.